refactor: move loading state to weather context
