+++ src/components/loading-indicator.jsx
@@ -1,7 +1,7 @@
 import LoadingIcon from "public/icons/loading.svg";
 
 const LoadingIndicator = () => (
-    <LoadingIcon className="animate-spin -ml-1 mr-3 h-5 w-5 fill-current" />
+    <LoadingIcon className="animate-spin h-5 w-5 fill-current ml-2" />
 );
 
 export default LoadingIndicator;
