+++ src/components/weather.jsx
@@ -5,7 +5,7 @@ import WeatherForecastItem from "./weather-forecast-item";
 import WeatherForecastList from "./weather-forecast-list";
 
 const Weather = () => {
-    const { forecast, loading } = useContext(WeatherContext);
+    const { forecast } = useContext(WeatherContext);
 
     return (
         <div className="w-full max-w-[572px] mx-auto">
@@ -15,7 +15,6 @@ const Weather = () => {
                     forecast.map((data, i) => (
                         <WeatherForecastItem
                             data={data}
-                            loading={loading}
                             index={i}
                             key={`weather_item_${i}`}
                             active={i === 0}
